backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
httpx
python-jose[cryptography]   # for JWT (or use PyJWT)
passlib[bcrypt]
python-multipart
aiofiles
scikit-learn
numpy
apscheduler
